save 4.27 修改 控件编辑 逻辑

+++ app/src/main/java/com/wuruoye/know/ui/edit/vm/RecordTypeEditViewModel.kt
@@ -131,7 +131,9 @@ class RecordTypeEditViewModel(
                 }
             }
 
-            recordViewDao.delete(RecordTypeSelect.getType(view), view.id!!)
+            if (view.id != null) {
+                recordViewDao.delete(RecordTypeSelect.getType(view), view.id!!)
+            }
         }
     }
